+++ client/src/extension.ts
@@ -1,28 +1,29 @@
 /* --------------------------------------------------------------------------------------------
  * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
+ * Licensed under the MIT License. See LICENSE in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import * as path from 'path';
-import { workspace, ExtensionContext } from 'vscode';
+import * as path from "path";
+import { workspace, ExtensionContext } from "vscode";
 
 import {
 	LanguageClient,
 	LanguageClientOptions,
 	ServerOptions,
 	TransportKind
-} from 'vscode-languageclient';
+} from "vscode-languageclient";
+import { defaultSettings, cssLanguages } from "./constants";
 
 let client: LanguageClient;
 
 export function activate(context: ExtensionContext) {
 	// The server is implemented in node
 	let serverModule = context.asAbsolutePath(
-		path.join('server', 'out', 'server.js')
+		path.join("server", "out", "server.js")
 	);
 	// The debug options for the server
-	// --inspect=6009: runs the server in Node's Inspector mode so VS Code can attach to the server for debugging
-	let debugOptions = { execArgv: ['--nolazy', '--inspect=6009'] };
+	// --inspect=6009: runs the server in Node"s Inspector mode so VS Code can attach to the server for debugging
+	let debugOptions = { execArgv: ["--nolazy", "--inspect=6009"] };
 
 	// If the extension is launched in debug mode then the debug server options are used
 	// Otherwise the run options are used
@@ -35,20 +36,33 @@ export function activate(context: ExtensionContext) {
 		}
 	};
 
+	let languages: string[] = defaultSettings.languages;
+	try {
+		const setting = workspace.getConfiguration("jsonColorToken");
+		languages = setting.get("languages");
+	} catch (err) {
+		console.warn("Unable to get configurations!");
+	}
 	// Options to control the language client
 	let clientOptions: LanguageClientOptions = {
-		// Register the server for plain text documents
-		documentSelector: [{ scheme: 'file', language: 'json' }, { scheme: 'file', language: 'css' }, { scheme: 'file', language: 'less' }],
+		// Register the server for json, jsonc files (color token detection)
+		// and css files for color token variable preview
+		documentSelector: [...languages, ...cssLanguages]
+			.map((lang) => {
+				return {
+					scheme: "file", language: lang
+				};
+			}),
 		synchronize: {
-			// Notify the server about file changes to '.clientrc files contained in the workspace
-			fileEvents: workspace.createFileSystemWatcher('**/.clientrc')
+			// Notify the server about file changes to ".clientrc files contained in the workspace
+			fileEvents: workspace.createFileSystemWatcher("**/.clientrc")
 		}
 	};
 
 	// Create the language client and start the client.
 	client = new LanguageClient(
-		'jsonColorToken',
-		'JSON Color Token LSP',
+		"jsonColorToken",
+		"JSON Color Token LSP",
 		serverOptions,
 		clientOptions
 	);