+++ server/src/server.ts
@@ -1,9 +1,9 @@
 /* --------------------------------------------------------------------------------------------
  * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
+ * Licensed under the MIT License. See LICENSE in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
- import {
+import {
 	createConnection,
 	TextDocuments,
 	ProposedFeatures,
@@ -18,21 +18,35 @@
 	ColorPresentation,
 	DidChangeTextDocumentParams,
 	TextDocumentChangeEvent,
-	DidChangeConfigurationParams
-} from 'vscode-languageserver';
-
+	DidChangeConfigurationParams,
+	DefinitionParams,
+	Definition,
+	Range
+} from "vscode-languageserver";
 import {
 	TextDocument
-} from 'vscode-languageserver-textdocument';
-import { IColors } from './IColors';
+} from "vscode-languageserver-textdocument";
+import { IColors } from "./IColors";
+import {
+	JSONColorTokenSettings,
+	defaultSettings,
+	cssLanguages,
+	colorTokenPattern,
+	cssVariablePattern,
+	jsonKeyPattern
+} from "./constants";
 
-// Create a connection for the server, using Node's IPC as a transport.
+// Create a connection for the server, using Node"s IPC as a transport.
 // Also include all preview / proposed LSP features.
 let connection = createConnection(ProposedFeatures.all);
 
 // Create a simple text document manager. 
 let documents: TextDocuments<TextDocument> = new TextDocuments(TextDocument);
-let jsonColorTokenCache: { [documentUri: string]: { [variable: string]: string} } = {};
+let colorTokenCache: {
+	[documentUri: string]: {
+		[variable: string]: { color: string, range: Range }
+	}
+} = {};
 
 let hasConfigurationCapability: boolean = false;
 let hasWorkspaceFolderCapability: boolean = false;
@@ -52,7 +66,8 @@ connection.onInitialize((params: InitializeParams) => {
 	const result: InitializeResult = {
 		capabilities: {
 			textDocumentSync: TextDocumentSyncKind.Incremental,
-			colorProvider: true
+			colorProvider: true,
+			definitionProvider: true
 		}
 	};
 	if (hasWorkspaceFolderCapability) {
@@ -72,28 +87,24 @@ connection.onInitialized(() => {
 	}
 	if (hasWorkspaceFolderCapability) {
 		connection.workspace.onDidChangeWorkspaceFolders(_event => {
-			connection.console.log('Workspace folder change event received.');
+			connection.console.log("Workspace folder change event received.");
 		});
 	}
 });
 
-// The example settings
-interface JSONColorTokenSettings {
-	maxNumberOfColorTokens: number;
-	colorTokenCasing: "Uppercase" | "Lowercase";
-}
-
-// The global settings, used when the `workspace/configuration` request is not supported by the client.
-// Please note that this is not the case when using this server with the client provided in this example
-// but could happen with other clients.
-const defaultSettings: JSONColorTokenSettings = { 
-	maxNumberOfColorTokens: 1000,
-	colorTokenCasing: "Uppercase"
-};
 let globalSettings: JSONColorTokenSettings = defaultSettings;
 
 // Cache the settings of all open documents
 let documentSettings: Map<string, Thenable<JSONColorTokenSettings>> = new Map();
+let cachedLanguages: string[] | undefined = undefined;
+
+async function isLanguageIncludedForColorTokenDetection(languageId: string): Promise<boolean> {
+	if (!cachedLanguages) {
+		const settings = await getRemoteConfiguration() as JSONColorTokenSettings;
+		cachedLanguages = settings.languages;
+	}
+	return cssLanguages.indexOf(languageId) < 0 && cachedLanguages.indexOf(languageId) >= 0;
+}
 
 connection.onDidChangeConfiguration((change: DidChangeConfigurationParams) => {
 	if (hasConfigurationCapability) {
@@ -118,7 +129,7 @@ connection.onDidChangeTextDocument((change: DidChangeTextDocumentParams) => {
 	if (!!document) {
 		findColorTokens(document);
 	}
-	documents.all().forEach(updateJsonColorTokenCache);
+	documents.all().forEach(updateColorTokenCache);
 });
 
 function getDocumentSettings(resource: string): Thenable<JSONColorTokenSettings> {
@@ -127,19 +138,23 @@ function getDocumentSettings(resource: string): Thenable<JSONColorTokenSettings>
 	}
 	let result = documentSettings.get(resource);
 	if (!result) {
-		result = connection.workspace.getConfiguration({
-			scopeUri: resource,
-			section: 'jsonColorToken'
-		});
+		result = getRemoteConfiguration(resource);
 		documentSettings.set(resource, result);
 	}
 	return result;
 }
 
+function getRemoteConfiguration(resource?: string): Promise<any> {
+	return connection.workspace.getConfiguration({
+		scopeUri: resource,
+		section: "jsonColorToken"
+	});
+}
+
 // Only keep settings for open documents
 documents.onDidClose(e => {
-	if (!!jsonColorTokenCache[e.document.uri]) {
-		delete jsonColorTokenCache[e.document.uri];
+	if (!!colorTokenCache[e.document.uri]) {
+		delete colorTokenCache[e.document.uri];
 	}
 	documentSettings.delete(e.document.uri);
 });
@@ -147,13 +162,9 @@ documents.onDidClose(e => {
 // The content of a text document has changed. This event is emitted
 // when the text document first opened or when its content has changed.
 documents.onDidChangeContent(async (change: TextDocumentChangeEvent<TextDocument>) => {
-	updateJsonColorTokenCache(change.document);
+	updateColorTokenCache(change.document);
 	await findColorTokens(change.document);
 });
-
-// A color token can be with/without opacity (last two digits)
-// e.g. without opacity #000000, with opacity #00000050
-const colorTokenPattern = /#[0-9a-fA-F]{6}([0-9]{2})?/g;
 
 function isColorToken(token: string | number | undefined): boolean {
 	if (typeof token === "string") {
@@ -163,18 +174,33 @@ function isColorToken(token: string | number | undefined): boolean {
 	return false;
 }
 
-async function updateJsonColorTokenCache(textDocument: TextDocument): Promise<void> {
-	if (textDocument.languageId === "json") {
+async function updateColorTokenCache(textDocument: TextDocument): Promise<void> {
+	if (await isLanguageIncludedForColorTokenDetection(textDocument.languageId)) {
 		let text = textDocument.getText();
 		try {
 			let jsonObj = JSON.parse(text);
-			let colorTokenObj: {[variable: string]: string} = {};
-			for (let key in jsonObj) {
-				if (isColorToken(jsonObj[key])) {
-					colorTokenObj[key] = jsonObj[key];
+			let colorTokenObj: {
+				[variable: string]: {
+					color: string,
+					range: Range
+				}
+			} = {};
+			const regex = new RegExp(jsonKeyPattern);
+			let m: RegExpExecArray | null;
+			while ((m = regex.exec(text))) {
+				const variableName = m.groups?.varDoulbeQuote ?? m.groups?.varSingleQuote;
+				if (!!variableName && isColorToken(jsonObj[variableName])) {
+					colorTokenObj[variableName] = {
+						color: jsonObj[variableName],
+						range: {
+							start: textDocument.positionAt(m.index),
+							end: textDocument.positionAt(m.index + variableName.length)
+						}
+					}
 				}
 			}
-			jsonColorTokenCache[textDocument.uri] = colorTokenObj;
+
+			colorTokenCache[textDocument.uri] = colorTokenObj;
 		} catch (error) {
 			// Swallow the error
 		}
@@ -183,51 +209,52 @@ async function updateJsonColorTokenCache(textDocument: TextDocument): Promise<vo
 
 let colors: IColors[] = [];
 async function findColorTokens(textDocument: TextDocument): Promise<void> {
-	let text = textDocument.getText();
-	if (textDocument.languageId === "json") {
-		// In this simple example we get the settings for every validate run.
-		let settings = await getDocumentSettings(textDocument.uri);
-		
+	const text = textDocument.getText();
+	// Get the maxNumberOfColorTokens for every run.
+	const settings = await getDocumentSettings(textDocument.uri);
+	const { maxNumberOfColorTokens } = settings;
+
+	if (await isLanguageIncludedForColorTokenDetection(textDocument.languageId)) {
 		let regex = new RegExp(colorTokenPattern);
 		let m: RegExpExecArray | null;
 		colors = [];
 		let numTokens = 0;
-	
-		while ((m = regex.exec(text)) && numTokens < settings.maxNumberOfColorTokens) {
+
+		while ((m = regex.exec(text)) && numTokens < maxNumberOfColorTokens) {
 			numTokens++;
-			
+
 			colors.push({
 				range: {
 					start: textDocument.positionAt(m.index),
 					end: textDocument.positionAt(m.index + m[0].length)
-				}, 
+				},
 				color: m[0]
 			});
 		}
-	} else if (textDocument.languageId === "css" || textDocument.languageId === "less") {
-		let pattern = /var\(--([a-zA-Z0-9\-]+)\)/g;
+	} else if (cssLanguages.indexOf(textDocument.languageId) >= 0) {
+		// Get reference to style variables
+		const regex = new RegExp(cssVariablePattern);
 		let m: RegExpExecArray | null;
 		colors = [];
-		while ((m = pattern.exec(text))) {
-			for (let documentUri in jsonColorTokenCache) {
-				let variableName = m[1];
-				if (jsonColorTokenCache[documentUri][variableName]) {
-					colors.push({
-						range: {
-							start: textDocument.positionAt(m.index),
-							end: textDocument.positionAt(m.index + m[0].length)
-						}, 
-						color: jsonColorTokenCache[documentUri][variableName]
-					});
-					break;
-				}
+		while ((m = regex.exec(text))) {
+			let variableName = m.groups?.cssVar || "";
+			// Show color preview of the first matching color token in color token cache
+			const documentUri = Object.keys(colorTokenCache).find((uri) => !!colorTokenCache[uri][variableName]);
+			if (!!documentUri) {
+				colors.push({
+					range: {
+						start: textDocument.positionAt(m.index),
+						end: textDocument.positionAt(m.index + m[0].length)
+					},
+					color: colorTokenCache[documentUri][variableName].color
+				});
 			}
 		}
 	}
 }
 
-function parseColor (color: string): Color {
-	const red = parseInt(color.slice(1, 3), 16)/ 255;
+function parseColor(color: string): Color {
+	const red = parseInt(color.slice(1, 3), 16) / 255;
 	const green = parseInt(color.slice(3, 5), 16) / 255;
 	const blue = parseInt(color.slice(5, 7), 16) / 255;
 	let alpha = 1.0;
@@ -260,6 +287,36 @@ function stringifyColor(color: Color, casing: "Uppercase" | "Lowercase"): string
 	return result;
 }
 
+// For color token variables referenced in css/less files,
+// go to definition will bring to the json file where the token is defined.
+connection.onDefinition((params: DefinitionParams): Definition | undefined => {
+	const { textDocument, position } = params;
+	const textDocumentObj = documents.get(textDocument.uri);
+	if (!textDocumentObj) {
+		return;
+	}
+	const languageID = textDocumentObj.languageId;
+	if (cssLanguages.indexOf(languageID) >= 0) {
+		// Get the line of the text and try to parse out the first referenced css variable, if any
+		const regex = new RegExp(cssVariablePattern);
+		const lineText = textDocumentObj.getText({
+			start: { line: position.line, character: 0 },
+			end: { line: position.line + 1, character: 0 }
+		});
+		const m = regex.exec(lineText);
+		if (!!m) {
+			const variableName = m.groups?.cssVar || "";
+			const documentUris = Object.keys(colorTokenCache).filter((uri) => !!colorTokenCache[uri][variableName]);
+			return documentUris.map((uri) => {
+				return {
+					uri: uri,
+					range: colorTokenCache[uri][variableName].range
+				};
+			});
+		}
+	}
+});
+
 connection.onDocumentColor(async (params: DocumentColorParams): Promise<ColorInformation[]> => {
 	const document = documents.get(params.textDocument.uri);
 	if (!document) {
@@ -278,7 +335,7 @@ connection.onDocumentColor(async (params: DocumentColorParams): Promise<ColorInf
 
 connection.onColorPresentation(async (params: ColorPresentationParams): Promise<ColorPresentation[]> => {
 	const document = documents.get(params.textDocument.uri);
-	if (!!document && document.languageId === "json") {
+	if (!!document && await isLanguageIncludedForColorTokenDetection(document.languageId)) {
 		let settings = await getDocumentSettings(params.textDocument.uri);
 		return [{ label: stringifyColor(params.color, settings.colorTokenCasing) }];
 	}